Implement outer/inner drop region indicator
There are two drop regions per direction.

The inner region is closer to the center of the window. Dropping on
this region will split the target container and put the container within the
split at the given direction beside the target container.

The outer region is close to the edge (currently 30px from the edge).
This will place the container as a sibling in the given direction
within the parent container.

The sibling (outer region) drop type is not yet implemented. This
commit only covers the indicator.