Handle a case of a soon to be gone window calling XSetInputFocus
This is a very rare case happening mostly for emacs-helm users.  It
can be reproduced without any special emacs configuration using
the x-focus-frame function and then immediately quitting emacs.

The problem was that in this case emacs calls XSetInputFocus with
revert to parent setting, which afterwards at the exit of emacs causes
keyboard focus to be lost and not set to the ewmh support window.