i3bar: make better use of compile time string length evaluation
The i3bar program contains a couple of locations where we use strlen on
a string that clearly is known at compile time and contained in the
binary. There is no point in doing so when the same information can be
retrieved statically by means of the sizeof operator.
This change fixes those occurrences. It also replaces a couple of magic
lengths with their symbolic counterparts and makes use of the return
value provided by the *printf* family of functions to avoid duplicated
string length assessment.