If no trayoutput matches, take first
When changing the screen, the tray was searching for a matching output.
But if no screen matched, it neglected the setting and didn't show
anything.

As the default, the i3bar shows a tray on the first matching output.
Its intention is to display a tray. So it should also show a tray, if
the monitors don't match.

Fixes #3317