Set WM_STATE to normal when focusing a window
Some applications request iconic state when they lose focus. One notable
example is games running in Wine that use the Windows exclusive
full-screen API.
While this is perfectly sound behavior on a floating window manager, we
do not support the concept of iconized windows, so not only do we not
handle this case, but there is also no sane way to escape that state.
This leaves the aforementioned games in a paused state, with the
contents reduced to a tiny corner of the window, and with no audio
playing.
The workaround, up until now, was to enable Wine's virtual desktop, but
that is pretty inconvenient to use.
To remedy this, reset WM_STATE to a normal state when the window is
focused. This allows the application to do whatever it wants to do while
it's not focused (pause the game, let the GPU idle, etc), and avoids
breaking it with a simple workspace switch.