commands: (rename) check workspace number
Assuming two workspaces 1 and 2, renaming the workspace 2 to "1: foo"
will reassign the number 1 already in use to the new workspace and break
the "workspace next_on_output" command.

This patch fixes this behavior by checking if a different workspace with
the same number does not already exist when renaming a given workspace.