Introduce layout_fill_order to Con
The layout fill order allows to configure how Cons are attached to a new
layout. The value 'default' keeps the traditional order, filling the
layout from left to right, or top lo bottom depending on horizontal or
vertical orientation. 'reverse' in turn means to fill a layout from
right to left, or bottom to top.

This commit does not yet add the behaviour described above, but merely
extends data structures to accommodate for this new option.

We only store a single fill order per Con and not any additional values
for the workspace layout and last split layout as those variables are
necessary to deal with the default layout which does not apply to the
fill order.