i3-config-wizard: Remove a pointless flush
I have no idea why this called xcb_flush(). The next thing that it does
is sending a GrabKeyboard request and waiting for its reply. This
already implicitly flushes the connection.

Also, later on, before waiting for events, the connection is flushed
again.

So... since I do not see the point of this flush, remove it.

Signed-off-by: Uli Schlachter <psychon@znc.in>