Call xcb_aux_sync() after IPC commands
I just got three test suite failures:

   Failed test 'Focus changed'
   at /usr/src/i3/distbuild/meson-private/dist-unpack/i3-4.19.2/testcases/t/003-ipc.t line 35.
          got: '6291467'
     expected: anything else
   Failed test 'left window focused'
   at /usr/src/i3/distbuild/meson-private/dist-unpack/i3-4.19.2/testcases/t/156-fullscreen-focus.t line 50.
          got: '6291467'
     expected: '4194304'
 left = 4194304
 right = 4194306
 third = 4194307
   Failed test '_NET_CURRENT_DESKTOP should be updated'
   at /usr/src/i3/distbuild/meson-private/dist-unpack/i3-4.19.2/testcases/t/294-update-ewmh-atoms.t line 104.
          got: '2'
     expected: '1'

My *guess* is that there is a race somewhere and the test suite expects
i3 to only answer to IPC commands after the X11 server already handled
the command. If so, then this was already previously racy: xcb_flush()
only ensures that everything was sent, but it does not guarantee that
the X11 server already handled the requests.

Thus, this adds a call to xcb_aux_sync() between handling an IPC command
and sending its reply. This ensures that the X11 server actually handled
everything related to that command already.

Signed-off-by: Uli Schlachter <psychon@znc.in>