Move more state updates to the new helper
This commit moves two more cases to the regrab_keys() helper function.

This is not just a straight forward refactoring since the order of
things is different for RG_LOAD_KEYMAP_AND_TRANSLATE before and after
this commit. While I do not think this makes much of a difference, I
still did this in its extra commit for bisectability.

Signed-off-by: Uli Schlachter <psychon@znc.in>