Move keyboard updates to an idle watcher
Instead of updating the keyboard state and regrabbing all the keys
immediately when requested, this simply starts an ev idle watcher. This
has the effect of batching updates: When multiple updates are requested
during a main loop iteration, only one update is actually done.

This fixes the new test case that I just added.

This might or might not fix https://github.com/i3/i3/issues/3924.

Signed-off-by: Uli Schlachter <psychon@znc.in>