draw_util_surface_free(): Remove unnecessary argument
Cairo allows querying the underlying xcb_connection_t* of a cairo
surface and thus this function does not really need the caller to pass
this in.

No idea if this change is really worth it. I am mostly doing this
because I can / because it seemed sensible (this means the caller cannot
pass different xcb_connection_t* to init and free, but that's likely a
very low risk).

Signed-off-by: Uli Schlachter <psychon@znc.in>