Add a "multi surface" mechanism
A "multi surface" (I am bad at naming things) consists of multiple
surface. Imagine you have a desk (the thing made out of wood) and you
arrange multiple sheets of papers on it. Each paper has a position and a
size. You can then "draw to the desk" and some of your drawing will end
up on some of the papers.

This is exactly what I mean with a "multi surface": It consists of up to
four surface_t instances. Each of them has a position which is its
offset from (0, 0). When drawing to the "multi surface", the drawing is
done on each of the contained surfaces, but with the offset applied.

All of the functions from draw_util are also available on "multi
surfaces".

A bit of a special case is multi_surface_copy_surface(). Compared to
draw_util_copy_surface(), there is an extra "fill_color" argument. All
areas that are not covered by any of the surfaces will be filled with
this fill color. Think of it as the background color of the desk in my
introductory example above (and also: No drawing to the desk is
possible; it always keeps its background color).

Signed-off-by: Uli Schlachter <psychon@znc.in>